requirements.txt
pandas==2.2.2
sqlalchemy==2.0.31
redis==5.0.7
cassandra-driver==3.29.1
scikit-learn==1.5.0
psycopg2-binary==2.9.9
streamlit==1.36.0